\name{PGS-package}
\alias{PGS-package}
\alias{PGS}
\docType{package}
\title{
\packageTitle{PGS}
}
\description{
\packageDescription{PGS}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{PGS}
\packageIndices{PGS}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{PGS}

Maintainer: \packageMaintainer{PGS}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{

}
