% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgsfit.obj.R
\name{pgsfit.obj}
\alias{coef.pgsfit.obj}
\alias{pgsfit.obj}
\alias{plot.pgsfit.obj}
\alias{print.pgsfit.obj}
\title{PGS S3 Class and Methods}
\usage{
pgsfit.obj(grid.err, lam.sel.vect, beta.shrink.corr.list, var.sand.corr.list,
  hat.R.list, flag.stop.corr.vect, iter.n.corr.vect, best.ind, bestall.ind, eps,
  iter.n, Pm.vect, lam.vect, scale.info, sis.dn, ConvergeMessage)

\method{print}{pgsfit.obj}(pgsfit.obj)

\method{plot}{pgsfit.obj}(pgsfit.obj, IQR.times = 1.5, text.size = 4,
  xlab.size = 14, ylab.size = 14)

\method{coef}{pgsfit.obj}(pgsfit.obj, pm.ind = NULL, p.threshold = 0.05,
  nonzero = TRUE)
}
\description{
\code{pgsfit.obj} is a S3 class to store PGS fitting results. Internally used by \code{pgsfit}. \code{\link{pgsfit.obj}} contains:
\itemize{
  \item{\code{grid.err}: cross-validation error grid}
  \item{\code{lam.sel.vect}: vector of selected lambda for each Pm}
  \item{\code{beta.shrink}: list of beta after shrinkage for each Pm}
  \item{\code{var.sand}: list of sandwich variance of beta for each Pm}
  \item{\code{hat.R}: list of estimated working correlation matrix for each Pm}
  \item{\code{convergenceError}: convergence error when iteration stopped for each Pm}
  \item{\code{iterationNumber}: iteration times when converged for each Pm}
  \item{\code{which.best}: index of the best Pm and lambda among convergent results}
  \item{\code{which.bestGlobal}: index of the best Pm and lambda among all results (including non-convergent results)}
  \item{\code{convergenceThreshold}: threshold of convergence error (i.e. \code{eps})}
  \item{\code{maxIteration}: maximum iteration number allowed (i.e. \code{iter.n})}
  \item{\code{Pm.vect}: working vector of tunning parameter Pm}
  \item{\code{lam.vect}: working vector of tunning parameter lambda}
  \item{\code{scale.info}: scaling information of y.vect, M, and COV to transfer beta estimate back to original scale}
  \item{\code{sis.name}: name (id) of genomic mark from sure independent screening results, truncated at \code{pm.max}}
  \item{\code{convergeMessage}: "Converged!", "Not converged!", or "Conditionally converged!}
}
}
\examples{
## Print PGS object:
pgsfit.obj

## Plot PGS object (a heat map visualizes the grid search errors):
plot(pgsfit.obj)

#Other parameters:
#IQR.times: cross-validation errors greater than IQR.times * IQR of the errors will be replaced by "NA" (represented by grey blocks). Default = 1.5.
#text.size: size of the text in grid. Default = 4.
#xlab.size: size of the x-axis labels. Default = 14.
#ylab.size: size of the y-axis labels. Default = 14.

## Return coefficients from PGS object:
coef(pgsfit.obj)

#Other parameters:
#whcih: an interger between 1 to pm.n specifying results at which Pm level to be returned. Defaul = NULL. If NULL, return the best results.
#p.threshold: threshold of p-values to filter out non-significant variables. Default = 0.05.
#nonzero: logical. If TRUE, only variables with non-zero beta results are returned. Default = TRUE.

#For more information, please visit: https://github.com/YinanZheng/PGS/wiki/Example:-miRNA-expression-and-lung-function
}
\seealso{
see \code{\link{pgsfit}} to run PGS algorithm and obtain \code{pgsfit.obj} object.
}

