% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotGrid.R
\name{plotGrid}
\alias{plotGrid}
\title{Visualize the Cross-Validation Error Grid with Heat Map}
\usage{
plotGrid(pgsobj, colrange = NULL)
}
\arguments{
\item{pgsobj}{a \code{pgsfit} object.}

\item{colrange}{a vector of color strings specifying the color range for heat map. The first color will be used for the lowest CV error; the second color will be used for the highest CV error. If not specified, \code{plotGrid} will generate dark to light purple color scheme by default.}
}
\description{
\code{plotGrid} is used to visualize the cross-validation (CV) error grid using a heat map.
}
\examples{
PGSfit = pgsfit(y.vect, id.vect, M, COV, preRank.vect, Pm.vect, lam.vect, seed = 1)
plotGrid(PGSfit)
}
\seealso{
see \code{\link{pgsfit}} to run PGS and obtain \code{pgsobj} object.
}

