% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/preRank.R
\name{preRank}
\alias{preRank}
\title{Site-by-site Independent Pre-ranking for Genomic marks}
\usage{
preRank(y.vect, id.vect, M, COV = NULL, method = c("LMM", "GEE"),
  rankby = c("coef", "p"), parallel = TRUE, ncore = detectCores(),
  write = FALSE)
}
\arguments{
\item{y.vect}{a vector of dependent variable.}

\item{id.vect}{a vector of subjuect ID.}

\item{M}{a data frame or matrix of genomic dataset. Rows represent samples, columns represent genomic marks.}

\item{COV}{a data frame or matrix of covariates dataset.}

\item{method}{a character string specifying fitting method. Linear mixed-effect model (\code{"LMM"}) and generalized estimation equation (\code{"GEE"}) method are supported. Default = \code{"LMM"}.}

\item{rankby}{a character string specifying how the geneomic marks should be ranked. They can be ranked by either their absolute value of coefficients (\code{"coef"}), or their statistics (t-statistics or Z-statistics) (\code{"stat"}). Default = \code{"coef"}.}

\item{parallel}{logical. Enable parallel computing feature? Default = \code{TRUE}.}

\item{ncore}{number of cores to run parallel. Effective when paralle = \code{TRUE}. By default, max number of cores will be used.}

\item{write}{logical. Export ranking results to csv file in the working directory if \code{TRUE}. Defaul = \code{FALSE}.}
}
\value{
A vector of ranked genomic marks.
}
\description{
\code{preRank} is used to fit mixed-effect models or generalize estimation equations across genomic marks and rank by coefficients or statistics.
}
\examples{
### Dataset preview
BJdata()

### Convert binary variables into factor type.
BJlung$gender = factor(BJlung$gender)
BJlung$heat = factor(BJlung$heat)
BJlung$cigwear = factor(BJlung$cigwear)

### Merge miRNA and lung function dataset.
BJdata <- merge(BJmirna, BJlung, by=c("SID","WD"))

### Data must be sorted by study subject ID and multiple measurements indicator.
BJdata <- BJdata[with(BJdata, order(SID, WD)), ]

### Extract dependent variable (lung function)
y.vect<-BJdata$FEV1

### Extract subjuect ID variable indicating repeated measures.
id.vect<-BJdata$SID

### Extract microRNA data matrix.
M<-BJdata[,3:168]

### Extract covariate data matrix.
COV<-BJdata[,170:179]

### LMM site-by-site pre-ranking results.
prerank_LMM_par = preRank(y.vect, id.vect, M, COV, method = "LMM")
### GEE site-by-site pre-ranking results.
prerank_GEE_par = preRank(y.vect, id.vect, M, COV, method = "GEE")
### Save the details in csv file
prerank_LMM_par = preRank(y.vect, id.vect, M, COV, method = "LMM", write = T)
}
\seealso{
see \code{\link{pgsfit}} using the results from \code{preRank} as input to run PGS.
}

