% Documentation file for PGS package
% Version: 0.2-0
% Authors: Kien Kieu & Marianne Mora
% License: CeCILL  
% Copyright © 2009 INRA, Université Paris Ouest Nanterre la Défense

\name{scaling}
\alias{scaling}
\title{Scale Figure, VecLat or FigLat objects}
\description{
  Scale (multiply by a real value) a subset.
}
\usage{
scaling(x, s)
}
\arguments{
  \item{x}{a \code{\link{Figure-class}}, \code{\link{VecLat-class}} or \code{\link{FigLat-class}} object.}
  \item{s}{the scaling factor.}
}
\value{
  An object of the same class as \code{x}.
}
\seealso{\code{\link{Figure-class}},
  \code{\link{VecLat-class}},
  \code{\link{FigLat-class}}.}
\keyword{spatial}
