% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sis.obj.R
\name{sis.obj}
\alias{coef.sis.obj}
\alias{plot.sis.obj}
\alias{print.sis.obj}
\alias{sis.obj}
\title{Sure Independent Screening S3 Class and Methods}
\usage{
sis.obj(name, estimate, stderr, p.value, bh.fdr, bonferroni.p, method)

\method{print}{sis.obj}(sis.obj)

\method{plot}{sis.obj}(sis.obj)

\method{coef}{sis.obj}(sis.obj)
}
\description{
\code{sis.obj} is a S3 class to store PGS sure independent screening results. Internally used by \code{sis} and \code{pgsfit}. \code{\link{sis.obj}} contains:
\itemize{
 \item{\code{name}: name (id) of variables}
 \item{\code{estimate}: beta estimate}
 \item{\code{stderr}: standard error}
 \item{\code{p.value}: raw p-value}
 \item{\code{bh.fdr}: Benjamini-Hochberg (BH) false discover rate (FDR)}
 \item{\code{bonferroni.p}: Bonferroni adjusted p-value}
 \item{\code{method}: independent screening method used in \code{sis}}
}
}
\examples{
## Print sure independent screening object:
sis.obj

## Plot sure independent screening object (histogram of p-values and Q-Q plot):
plot(sis.obj)

## Return sure independent screening results:
coefficients = coef(sis.obj)
head(coefficients)

#For more information, please visit: https://github.com/YinanZheng/PGS/wiki/Example:-miRNA-expression-and-lung-function
}
\seealso{
see \code{\link{sis}} to run sure independent screening procedures and obtain \code{sis.obj} object.
}

