\name{PGS-package}
\alias{PGS-package}
\alias{PGS}
\docType{package}
\title{
\packageTitle{PGS}
}
\description{
\packageDescription{PGS}
}
\details{
\code{\link{BJdata}} is used to load and preview PGS built-in example datasets \code{BJmirna} and \code{BJlung}.

\code{\link{pgsfit}} is used to fit and determine the best results from penalized GEE method across tunning parameter grid.

\code{\link{sis}} is used to conduct sure independent screening across high-dimensional variables.
}
\author{
\packageAuthor{PGS}

Maintainer: \packageMaintainer{PGS}
}
\references{
Zheng Y, Fei Z, Zhang W, Starren JB, Liu L, Baccarelli AA, Li Y, Hou L. PGS: a tool for association study of high-dimensional microRNA expression data with repeated measures. Bioinformatics. 2014;30(19):2802-7. doi: 10.1093/bioinformatics/btu396. PubMed PMID: 24947752; PubMed Central PMCID: PMC4173025. (\href{http://www.ncbi.nlm.nih.gov/pubmed/24947752}{pubmed})

Wang L, Zhou J, Qu A. Penalized generalized estimating equations for high-dimensional longitudinal data analysis. Biometrics. 2012;68(2):353-60. doi: 10.1111/j.1541-0420.2011.01678.x. PubMed PMID: 21955051. (\href{http://www.ncbi.nlm.nih.gov/pubmed/21955051}{pubmed})
}
\keyword{package}