% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGrid.R
\name{plotGrid}
\alias{plotGrid}
\title{Visualize the Cross-Validation Error Grid with Heat Map}
\usage{
plotGrid(pgsobj, IQR.times = 1.5, text.size = 4, xaxislab.size = 14,
  yaxislab.size = 14)
}
\arguments{
\item{pgsobj}{a \code{pgsfit} object.}

\item{IQR.times}{cross-validation errors greater than \code{IQR.times} * IQR of the errors will be replaced by "NA" (represented by grey blocks). Default = 1.5.}

\item{text.size}{size of the text in grid. Default = 4.}

\item{xaxislab.size}{size of the x-axis labels. Default = 14.}

\item{yaxislab.size}{size of the y-axis labels. Default = 14.}
}
\description{
\code{plotGrid} is used to visualize the cross-validation (CV) error grid using a heat map.
}
\examples{
PGSfit = pgsfit(y.vect, id.vect, M, COV, preRank_LMM_par, Pm.vect, lam.vect, seed = 1)
plotGrid(PGSfit)
}
\seealso{
see \code{\link{pgsfit}} to run PGS and obtain \code{pgsobj} object.
}

